requirements.txt
mip == 1.13.0; python_version >= '3.8'
pandas >= 1.1.0, < 1.3.0; python_version == '3.7'
pandas >= 1.3.0; python_version >= '3.8'

pyyaml >= 6.0
pathlib2 >= 2.3.7
